Added categories and a repository for them

+++ src/main/java/com/example/bookstore/domain/BookRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.CrudRepository;
 import java.util.List;
 
 public interface BookRepository extends CrudRepository<Book, Long> {
-		List<Book> findByid(int id);
+		List<Book> findByid(Long id);
 }
